opraveny chyby metody přesunuty do utils a metody mockovane
