+++ client/src/pages/NovaAkceFormik/NovaAkceFormik.tsx
@@ -2,46 +2,11 @@ import React, { useState } from "react";
 import { Formik, Form, Field } from "formik";
 import { Navigation } from "../../components/Navigation/Navigation";
 import "./styles.css";
-
-type FormValues = {
-  title: string;
-  misto?: string;
-  dates: string[];
-};
+import { handleSubmit } from "./utils";
 
 export const NovaAkceFormik: React.FC = () => {
   const [dates, setDates] = useState<string[]>([""]);
   const maxDates = 10;
-
-  const handleSubmit = async (values: FormValues) => {
-    const requestBody: Record<string, any> = {
-      name: values.title,
-      title: values.title,
-      dates: values.dates.map((date: string) => new Date(date).getTime()),
-    };
-
-    if (values.misto) {
-      requestBody.location = values.misto; // 🔹 Přidáme misto jen pokud je vyplněné
-    }
-
-    try {
-      const response = await fetch("/api/events", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify(requestBody),
-      });
-
-      if (!response.ok) {
-        throw new Error("Chyba při odesílání akce.");
-      }
-
-      alert("Akce byla úspěšně přidána!");
-    } catch (error) {
-      alert("Nastala chyba: " + error.message);
-    }
-  };
 
   return (
     <>
